Organised everything, added history, new Makefile, splitted main functions
