+++ .gitignore
@@ -1,0 +1,6 @@
+*.o
+*.a
+*.out
+./idea
+./vscode
+minishell
