+++ srcs/get_new_line.c
@@ -3,18 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   get_new_line.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: geymat <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: lcamerly <lcamerly@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/06 05:13:05 by geymat            #+#    #+#             */
-/*   Updated: 2024/03/06 05:29:38 by geymat           ###   ########.fr       */
+/*   Updated: 2024/03/07 04:06:35 by lcamerly         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
-#include "libft/libft.h"
-#include <stdio.h>
-#include <readline/readline.h>
-#include <readline/history.h>
+#include "../inc/minishell.h"
+
 
 static int	quotes(char     *str)
 {
@@ -45,11 +42,12 @@ char	*get_a_new_line(t_env *env)
 	char	*str;
 	char	*temp;
 
-	str = readline("minishell: ");
+	str = readline("MiniCheh -> ");
 	if (!str)
-		return (NULL);
+		return(write(1, "exit\n", 5), free(str), NULL);
+	add_history(str);
 	while (quotes(str))
-	{
+	{	
 		str = ft_strjoin_free_first(str, "\n");
 		if (!str)
 			return (NULL);
